Move RecordConvert to public API
